+++ fritzconnection/lib/fritztools.py
@@ -1,6 +1,31 @@
 """
 Some helper functions for more readable bits and bytes.
 """
+import math
+
+
+def byte_formatter(value):
+    """
+    Gets a large integer als value and returns a tuple with the value as
+    float and the matching dimension as string, i.e.
+    >>> byte_formatter(242981246)
+    (242.981246, 'MB')
+    Expects positive integer as input. Negative numbers are interpreted
+    as positive numbers. values < 1 are interpreted as 0.
+    """
+    dim = ['B', 'KB', 'MB', 'GB', 'TB']
+    value = abs(value)
+    if value < 1:
+        log = 0
+        num = 0
+    else:
+        log = min(int(math.log10(value) / 3), len(dim))
+        num = value / 1000 ** log
+    try:
+        dimension = dim[log]
+    except IndexError:
+        dimension = 'PB'
+    return num, dimension
 
 
 def format_num(num, unit='bytes'):
@@ -8,18 +33,10 @@ def format_num(num, unit='bytes'):
     Returns a human readable string of a byte-value.
     If 'num' is bits, set unit='bits'.
     """
-    if unit == 'bytes':
-        extension = 'B'
-    else:
-        # if it's not bytes, it's bits
-        extension = 'Bit'
-    for dimension in (unit, 'K', 'M', 'G', 'T'):
-        if num < 1024:
-            if dimension == unit:
-                return f'{num:3.1f} {dimension}'
-            return f'{num:3.1f} {dimension}{extension}'
-        num /= 1024
-    return f'{num:3.1f} P{extension}'
+    num, dim = byte_formatter(num)
+    if unit != 'bytes':
+        dim += 'it'  # then its Bit by default
+    return f'{num:3.1f} {dim}'
 
 
 def format_rate(num, unit='bytes'):
@@ -29,10 +46,11 @@ def format_rate(num, unit='bytes'):
     """
     return format_num(num, unit=unit) + '/s'
 
+
 def format_dB(num):
     """
     Returns a human readable string of dB. The value is divided
     by 10 to get first decimal digit
     """
     num /= 10
-    return f'{num:3.1f} {"dB"}' 
+    return f'{num:3.1f} {"dB"}'